requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-appconf==1.0.4
django-cleanup==5.2.0
django-crispy-forms==1.11.1
django-environ==0.4.5
django-heroku==0.3.1
django-imagekit==4.0.2
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.0
pilkit==2.0
Pillow==8.1.2
psycopg2==2.7.7
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
